Contact Preview splitted in bit-sized components
