+++ plugin-hrm-form/src/components/search/ContactPreview/index.js
@@ -1,0 +1,3 @@
+import ContactPreview from './ContactPreview';
+
+export default ContactPreview;
